+++ tests/test_all.nim
@@ -19,4 +19,5 @@ import t_correlator
 import t_renderer
 import t_cli_args
 import t_library_api
+import t_e2e_integration
 import test1